# backend/docs/swagger.yaml
basePath: /
definitions:
...
  handler.ErrorDetail:
    properties:
      code:
        example: NOT_FOUND
        type: string
      message:
        example: not found
        type: string
    type: object
  handler.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/handler.ErrorDetail'
    type: object
...
  handler.InviteAcceptResponse:
    properties:
      group_id:
        example: 665c1a2b3c4d5e6f7a8b9c0d
        type: string
      id:
        example: 665d1a2b3c4d5e6f7a8b9c0d
        type: string
      status:
        example: ACCEPTED
        type: string
    type: object
  handler.InviteDeclineResponse:
    properties:
      id:
        example: 665d1a2b3c4d5e6f7a8b9c0d
        type: string
      status:
        example: DECLINED
        type: string
    type: object
  handler.InviteListResponse:
    properties:
      invites:
        items: {}
        type: array
    type: object
...
  handler.MarkAllReadResponse:
    properties:
      updated:
        example: 3
        type: integer
    type: object
  handler.MarkReadResponse:
    properties:
      id:
        example: 665e1a2b3c4d5e6f7a8b9c0d
        type: string
      read:
        example: true
        type: boolean
    type: object
  handler.NotificationListResponse:
    properties:
      has_more:
        example: true
        type: boolean
      notifications:
        items: {}
        type: array
    type: object
...
  handler.SendInviteRequest:
    properties:
      user_id:
        example: 665b1a2b3c4d5e6f7a8b9c0d
        type: string
    type: object
  handler.UnreadCountResponse:
    properties:
      count:
        example: 3
        type: integer
    type: object
...
  model.Invite:
    properties:
      created_at:
        type: string
      group_id:
        type: string
      group_name:
        type: string
      id:
        type: string
      invited_by:
        type: string
      invited_by_name:
        type: string
      invitee_id:
        type: string
      status:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Social prediction betting API — create groups, place bets, resolve
    outcomes.
  title: youwont API
  version: "1.0"
paths:
...
  /groups/{id}/invites:
    post:
      consumes:
      - application/json
      description: Invite a user to a group. Sender must be a group member. Creates
        a notification for the invitee.
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      - description: User to invite
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.SendInviteRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Invite'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Send group invite
      tags:
      - invites
...
  /invites:
    get:
      description: Returns all pending group invites for the authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.InviteListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List my pending invites
      tags:
      - invites
  /invites/{id}/accept:
    post:
      description: Accept a pending group invite. Adds the user to the group and notifies
        the inviter.
      parameters:
      - description: Invite ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.InviteAcceptResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Accept an invite
      tags:
      - invites
  /invites/{id}/decline:
    post:
      description: Decline a pending group invite. The user can be re-invited later.
      parameters:
      - description: Invite ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.InviteDeclineResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Decline an invite
      tags:
      - invites
  /notifications:
    get:
      description: Returns the authenticated user's notification feed, newest first.
        Paginated.
      parameters:
      - default: 0
        description: Page number (0-indexed)
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page (max 50)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.NotificationListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List notifications
      tags:
      - notifications
  /notifications/{id}/read:
    post:
      description: Marks a single notification as read.
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.MarkReadResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Mark notification as read
      tags:
      - notifications
  /notifications/read-all:
    post:
      description: Marks all unread notifications as read for the authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.MarkAllReadResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Mark all notifications as read
      tags:
      - notifications
  /notifications/unread-count:
    get:
      description: Returns the number of unread notifications. Used for the bell badge.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.UnreadCountResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get unread notification count
      tags:
      - notifications
...
securityDefinitions:
  BearerAuth:
    description: Enter your Supabase JWT token with the "Bearer " prefix.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"